+++ ConsoleAppAids3/Properties/AssemblyInfo.cs
@@ -34,6 +34,12 @@
 
 	2018/11/26 7.0     DAG Eliminate the unreferenced system namespaces, and tag
                            my own assembly references with explanatory notes.
+
+	2019/01/30 7.1     DAG Reset the console row indicator when ScrollUp is
+	                       called on a FixedConsoleWriter object, so that the
+	                       object need not be re-created when you want to reuse
+	                       it further down screen. I'm calling this a point
+	                       release, although it is only a bug fix.
     ============================================================================
 */
 
@@ -49,7 +55,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration ( "" )]
 [assembly: AssemblyCompany ( "David A. Gray" )]
 [assembly: AssemblyProduct ( "WizardWrx .NET Utility Libraries" )]
-[assembly: AssemblyCopyright ( "Copyright © 2012-2018, David A. Gray" )]
+[assembly: AssemblyCopyright ( "Copyright © 2012-2019, David A. Gray" )]
 [assembly: AssemblyTrademark ( "This library is distributed under a three-clause BSD license." )]
 [assembly: AssemblyCulture ( "" )]
 
@@ -71,4 +77,4 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion ( "7.0.515.*" )]
+[assembly: AssemblyVersion ( "7.1.521.*" )]